Remove NexmoDeveloper::Application requirement
The rails engine logic is currently tied into NexmoDeveloper which means it’s not possible to use the engine in a normal rails app even though the code is there. Removing this logic allows us to use this in rails apps but also potentially removes the ability to use it without Rails. That assumption is currently untested. I unfortunately do not have time to do this nicely, I’m just hacking stuff apart so I can use this to render documentation.